Put the unit test back in test_check_leo.py. Kill the QApplication from the script!!
